+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/download/task/DownloadEpisodeTask.kt
@@ -8,10 +8,11 @@ import android.os.Build
 import android.system.ErrnoException
 import android.system.OsConstants
 import android.util.Log
+import android.widget.Toast
 import androidx.hilt.work.HiltWorker
+import androidx.work.CoroutineWorker
 import androidx.work.Data
 import androidx.work.ForegroundInfo
-import androidx.work.Worker
 import androidx.work.WorkerParameters
 import au.com.shiftyjelly.pocketcasts.analytics.EpisodeDownloadError
 import au.com.shiftyjelly.pocketcasts.models.entity.BaseEpisode
@@ -56,8 +57,10 @@ import kotlin.coroutines.resumeWithException
 import kotlin.coroutines.suspendCoroutine
 import kotlin.time.DurationUnit
 import kotlin.time.TimeSource
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.rx2.await
+import kotlinx.coroutines.withContext
 import okhttp3.Call
 import okhttp3.Callback
 import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
@@ -78,7 +81,7 @@ class DownloadEpisodeTask @AssistedInject constructor(
     var userEpisodeManager: UserEpisodeManager,
     @DownloadCallFactory private val callFactory: Call.Factory,
     @DownloadRequestBuilder private val requestBuilderProvider: Provider<Request.Builder>,
-) : Worker(context, params) {
+) : CoroutineWorker(context, params) {
 
     companion object {
         private const val MAX_RETRIES = 5
@@ -110,12 +113,14 @@ class DownloadEpisodeTask @AssistedInject constructor(
         const val OUTPUT_ERROR_MESSAGE = "error_message"
         const val OUTPUT_EPISODE_UUID = "episode_uuid"
         const val OUTPUT_CANCELLED = "cancelled"
+        const val FIRE_TOAST = "fire_toast"
     }
 
     private lateinit var episode: BaseEpisode
     private val episodeUUID: String? = inputData.getString(INPUT_EPISODE_UUID)
     private val pathToSaveTo: String? = inputData.getString(INPUT_PATH_TO_SAVE_TO)
     private val tempDownloadPath: String? = inputData.getString(INPUT_TEMP_PATH)
+    private val fireToast: Boolean = inputData.getBoolean(FIRE_TOAST, false)
 
     private var bytesDownloadedSoFar: Long = 0
     private var bytesRemaining: Long = 0
@@ -123,7 +128,7 @@ class DownloadEpisodeTask @AssistedInject constructor(
     private val timeSource = TimeSource.Monotonic
     private val startTimestamp = timeSource.markNow()
 
-    override fun doWork(): Result {
+    override suspend fun doWork(): Result {
         if (isStopped) {
             LogBuffer.i(LogBuffer.TAG_BACKGROUND_TASKS, "Cancelling execution of $episodeUUID download because we are already stopped")
             return Result.failure(failureData())
@@ -169,6 +174,11 @@ class DownloadEpisodeTask @AssistedInject constructor(
                     }
                 }
 
+                if (fireToast) {
+                    withContext(Dispatchers.Main) {
+                        Toast.makeText(context, context.getString(LR.string.download_completed_for_episode, episode.title), Toast.LENGTH_SHORT).show()
+                    }
+                }
                 LogBuffer.i(LogBuffer.TAG_BACKGROUND_TASKS, "Downloaded episode ${episode.title} ${episode.uuid}")
 
                 Result.success(outputData)