+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 7.67
 -----
-
+*   Bug Fixes
+    *   Fix deselected chapters not synced correctly between different platforms
+        ([#2357](https://github.com/Automattic/pocket-casts-android/pull/2357))
+    *   Adds download episode action button to shelf list in Now Playing
+        ([#2325](https://github.com/Automattic/pocket-casts-android/pull/2325))
 
 7.66
 -----